+++ himalaya/ridge/_random_search.py
@@ -2,13 +2,13 @@ import warnings
 import numbers
 
 import numpy as np
-from sklearn.model_selection import check_cv
 
 from ..backend import get_backend
 from ..backend._utils import _dtype_to_str
 from ..progress_bar import bar
 from ..scoring import l2_neg_loss
 from ..validation import check_random_state
+from ..validation import check_cv
 from ..kernel_ridge import generate_dirichlet_samples
 from ..kernel_ridge._random_search import _select_best_alphas
 
@@ -163,7 +163,7 @@ def solve_group_ridge_random_search(
     if n_alphas_batch is None:
         n_alphas_batch = len(alphas)
 
-    cv = check_cv(cv)
+    cv = check_cv(cv, Y)
     n_splits = cv.get_n_splits()
     for train, val in cv.split(Y):
         if len(val) == 0 or len(train) == 0: